requirements.txt
Django==1.8.5

PyMySql=0.6.7
django-filter==0.11.0
djangorestframework==3.2.4
mysql==0.0.1
mysql-connector-python==2.1.3
mysqlclient==1.3.6
pyreadline==2.1
six==1.10.0
python==3.4.0